worker/requirements.txt


psycopg[binary,pool]==3.2.3
redis==5.0.8
boto3==1.35.39
pypdf==5.0.1
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.3
openpyxl==3.1.5
openai==1.51.0
numpy==2.1.2
